+++ test/__init__.py
@@ -1,0 +1,4 @@
+# GP0 <-1k-> SCL
+# GP1 <-1k-> SDA
+# GP2 <-1k-> GP3
+# SCL, SDA -> EEPROM (0x50)