+++ docs/source/history.rst
@@ -16,7 +16,7 @@ ADC/DAC:
 
 I2C:
     * **Deprecated** *I2C_cancel* and *I2C_is_idle*. Bus is now managed automatically. You can use :any:`_i2c_release` and :any:`_i2c_status` as replacement if needed.
-    * Fixed. Low SCL and low SDA exceptions were swapped. Fixed.
+    * Fixed. Low SCL and low SDA exceptions were swapped.
     * When the I2C bus detects SDA activity, the next transfer does not work fine. Prevented via software. See :any:`_i2c_status`.
 
 Misc: